Fixes
Fixed bug in STM32 Code- we were sending 7 SPI clock bits instead of 8

Fixed logic in Verilog implementation of the SPI module. It is now asserting a signal when the register is written with 8 bits, which allows us to now transfer bytes to registers inside the CPLD. Exactly what we aimed for. Verilog is damn easier than VHDL.